requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
cycler==0.10.0
Cython==0.29.17
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.0
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
more-itertools==8.7.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.5
patsy==0.5.1
Pillow==8.2.0
plotly==4.14.3
pmdarima==1.8.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
Quandl==3.6.0
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.23
statsmodels==0.12.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
